refactor: split LensExplorer and CameraExplorer into smaller modules
LensExplorer (375 → 4 files, max 121):
- constants.ts — types, column defs, filter ranges
- LensFilters.tsx — filter dropdowns + chip groups
- LensResults.tsx — desktop table + mobile cards
- LensExplorer.tsx — orchestrator with state + filtering

CameraExplorer (326 → 4 files, max 115):
- constants.ts — types, column defs, filter ranges
- CameraFilters.tsx — filter dropdowns + chip groups
- CameraResults.tsx — desktop table + mobile cards
- CameraExplorer.tsx — orchestrator with state + filtering

Closes #289
